refactor: Estructura de Proyecto y Nomenclatura. Buenas Prácticas P1
