+++ src/controllers/driver.controller.mjs
@@ -1,6 +1,6 @@
-import { getAllByPosition, getByDriverId, save, deleteById, updateById } from "../../service/driver.service.mjs";
-import { ContentTypeException } from "../../service/exceptions/MyException.mjs";
-import { CONTENT_TYPE_EXP } from "../../utils/messages/error/headers.error.messages.mjs";
+import { getAllByPosition, getByDriverId, save, deleteById, updateById } from "../services/driver.service.mjs";
+import { ContentTypeException } from "../services/exceptions/MyException.mjs";
+import { CONTENT_TYPE_EXP } from "../utils/messages/error/headers.error.messages.mjs";
 
 const HEADER_TYPE = "Content-Type"
 const APPLICATION_CONTENT = "application/json"
