+++ src/services/exceptions/MyException.mjs
